Extract align rules tests related to Spring
