+++ src/integTest/groovy/nebula/plugin/resolutionrules/AlignRulesBasicWithCoreSpec.groovy
@@ -184,371 +184,6 @@ class AlignRulesBasicWithCoreSpec extends IntegrationTestKitSpec {
     }
 
     @Unroll
-    def 'spring boot 1.x dependencies interaction | with #versionType | core alignment #coreAlignment'() {
-        given:
-        buildFile.delete()
-        buildFile.createNewFile()
-        def extSpringVersion = '4.2.4.RELEASE'
-        def extSpringBootVersion = '1.5.6.RELEASE'
-
-        // in Spring Boot 1.x plugin, dependency management is added automatically
-        springBootBasedBuildFileWith(extSpringBootVersion, extSpringVersion, usesForce, requestedVersion, '')
-        buildFile << """
-            dependencies {
-                compile "org.springframework:spring-core$requestedVersion"
-                compile "org.springframework.boot:spring-boot-starter"
-                compile "org.springframework.boot:spring-boot-starter-web"
-            }
-            """.stripIndent()
-
-        new File("${projectDir}/gradle.properties").delete()
-
-        rulesJsonFile << alignSpringRule()
-
-        when:
-        def result = runTasks('dependencyInsight', '--dependency', 'org.springframework', "-Dnebula.features.coreAlignmentSupport=$coreAlignment")
-        def output = result.output
-
-        then:
-        dependencyInsightContains(output, 'org.springframework:spring-aop', resultingVersionForMost)
-        dependencyInsightContains(output, 'org.springframework:spring-beans', resultingVersionForMost)
-        dependencyInsightContains(output, 'org.springframework:spring-expression', resultingVersionForMost)
-
-        if (!resultsAreAligned) {
-            dependencyInsightContains(output, 'org.springframework:spring-core', extSpringVersion)
-        } else {
-            dependencyInsightContains(output, 'org.springframework:spring-core', resultingVersionForMost)
-        }
-
-        where:
-        versionType                   | coreAlignment | requestedVersion     | resultingVersionForMost | usesForce | resultsAreAligned
-        'provided version'            | false         | ''                   | '4.3.10.RELEASE'        | false     | true
-        'provided version'            | true          | ''                   | '4.3.10.RELEASE'        | false     | true
-
-        'declared version'            | false         | ':\${springVersion}' | '4.3.10.RELEASE'        | false     | true
-        'declared version'            | true          | ':\${springVersion}' | '4.3.10.RELEASE'        | false     | false
-
-        'declared version and forced' | false         | ':\${springVersion}' | '4.2.4.RELEASE'         | true      | true
-        'declared version and forced' | true          | ':\${springVersion}' | '4.3.10.RELEASE'        | true      | false
-    }
-
-    @Unroll
-    def 'spring boot 2.x dependencies interaction | with #versionType | core alignment #coreAlignment'() {
-        given:
-        buildFile.delete()
-        buildFile.createNewFile()
-        def extSpringVersion = '4.2.4.RELEASE'
-        def extSpringBootVersion = '2.1.4.RELEASE'
-
-        // in Spring Boot 2.x plugin, the `io.spring.dependency-management` plugin is added for dependency management
-        springBootBasedBuildFileWith(extSpringBootVersion, extSpringVersion, usesForce, requestedVersion, "\napply plugin: \"io.spring.dependency-management\"")
-        buildFile << """
-            dependencies {
-                compile "org.springframework:spring-core$requestedVersion"
-                compile "org.springframework.boot:spring-boot-starter"
-                compile "org.springframework.boot:spring-boot-starter-web"
-            }
-            """.stripIndent()
-
-        new File("${projectDir}/gradle.properties").delete()
-
-        rulesJsonFile << alignSpringRule()
-
-        when:
-        def result = runTasks('dependencyInsight', '--dependency', 'org.springframework', "-Dnebula.features.coreAlignmentSupport=$coreAlignment")
-        def output = result.output
-
-        then:
-        dependencyInsightContains(output, 'org.springframework:spring-aop', resultingVersionForMost)
-        dependencyInsightContains(output, 'org.springframework:spring-beans', resultingVersionForMost)
-        dependencyInsightContains(output, 'org.springframework:spring-expression', resultingVersionForMost)
-
-        if (!resultsAreAligned) {
-            dependencyInsightContains(output, 'org.springframework:spring-core', extSpringVersion)
-        } else {
-            dependencyInsightContains(output, 'org.springframework:spring-core', resultingVersionForMost)
-        }
-
-        where:
-        versionType                   | coreAlignment | requestedVersion     | resultingVersionForMost | usesForce | resultsAreAligned
-        'provided version'            | false         | ''                   | '5.1.6.RELEASE'         | false     | true
-        'provided version'            | true          | ''                   | '5.1.6.RELEASE'         | false     | true
-
-        'declared version'            | false         | ':\${springVersion}' | '5.1.6.RELEASE'         | false     | true
-        'declared version'            | true          | ':\${springVersion}' | '5.1.6.RELEASE'         | false     | false
-
-        'declared version and forced' | false         | ':\${springVersion}' | '4.2.4.RELEASE'         | true      | true
-        'declared version and forced' | true          | ':\${springVersion}' | '5.1.6.RELEASE'         | true      | false
-    }
-
-    @Unroll
-    def 'spring boot 2.x dependencies interaction without Spring dependency management | with #versionType | core alignment #coreAlignment'() {
-        given:
-        buildFile.delete()
-        buildFile.createNewFile()
-        def extSpringVersion = '4.2.4.RELEASE'
-        def extSpringBootVersion = '2.1.4.RELEASE'
-
-        // in Spring Boot 2.x plugin, the `io.spring.dependency-management` plugin is added for dependency management. We are not including it here.
-        springBootBasedBuildFileWith(extSpringBootVersion, extSpringVersion, usesForce, requestedVersion, '')
-        buildFile << """
-            dependencies {
-                compile "org.springframework:spring-core$requestedVersion"
-                compile "org.springframework.boot:spring-boot-starter:$extSpringBootVersion"
-                compile "org.springframework.boot:spring-boot-starter-web:$extSpringBootVersion"
-            }
-            """.stripIndent()
-
-        new File("${projectDir}/gradle.properties").delete()
-
-        rulesJsonFile << alignSpringRule()
-
-        when:
-        def result = runTasks('dependencyInsight', '--dependency', 'org.springframework', "-Dnebula.features.coreAlignmentSupport=$coreAlignment")
-        def output = result.output
-
-        then:
-        dependencyInsightContains(output, 'org.springframework:spring-aop', resultingVersionForMost)
-        dependencyInsightContains(output, 'org.springframework:spring-beans', resultingVersionForMost)
-        dependencyInsightContains(output, 'org.springframework:spring-expression', resultingVersionForMost)
-
-        // and results are aligned!
-        dependencyInsightContains(output, 'org.springframework:spring-core', resultingVersionForMost)
-
-        where:
-        versionType                   | coreAlignment | requestedVersion     | resultingVersionForMost | usesForce
-        'declared version'            | false         | ':\${springVersion}' | '5.1.6.RELEASE'         | false
-        'declared version'            | true          | ':\${springVersion}' | '5.1.6.RELEASE'         | false
-
-        'declared version and forced' | false         | ':\${springVersion}' | '4.2.4.RELEASE'         | true
-        'declared version and forced' | true          | ':\${springVersion}' | '4.2.4.RELEASE'         | true
-    }
-
-    @Unroll
-    def 'missing cases: statically defined dependency: alignment styles cannot resolve missing dependency when lower versions are missing | core alignment: #coreAlignment'() {
-        given:
-        def mavenrepo = createDepenendenciesForMissingVersionsTests()
-        new File("${projectDir}/gradle.properties").delete()
-
-        buildFile << """
-            repositories {
-                ${mavenrepo.mavenRepositoryBlock}
-            }
-            dependencies {
-                compile 'test.nebula:a:${definedVersion}'
-                compile 'test.nebula:b:${definedVersion}'
-                compile 'test.nebula:c:${definedVersion}'
-            }
-            """.stripIndent()
-
-        rulesJsonFile << alignTestNebulaRule()
-
-        when:
-        def result = runTasks(*tasks(coreAlignment))
-
-        then:
-        dependencyInsightContains(result.output, "test.nebula:a", ABResultingVersion)
-        dependencyInsightContains(result.output, "test.nebula:b", ABResultingVersion)
-        dependencyInsightContains(result.output, "test.nebula:c", CResultingVersion)
-
-        if (coreAlignment) {
-            // align on the non-failed version
-            assert result.output.contains("belongs to platform aligned-platform:rules-0-for-test.nebula-or-test.nebula.ext:$ABResultingVersion")
-        }
-
-        where:
-        definedVersionType | definedVersion | coreAlignment | ABResultingVersion | CResultingVersion
-        "static version"   | "0.5.0"        | false         | "0.5.0"            | "FAILED"
-        "static version"   | "0.5.0"        | true          | "0.5.0"            | "FAILED"
-    }
-
-    @Unroll
-    def 'missing cases: statically defined dependency: alignment styles cannot resolve missing dependency when higher versions are missing | core alignment: #coreAlignment'() {
-        def mavenrepo = createDepenendenciesForMissingVersionsTests()
-        new File("${projectDir}/gradle.properties").delete()
-
-        buildFile << """
-            repositories {
-                ${mavenrepo.mavenRepositoryBlock}
-            }
-            dependencies {
-                compile 'test.nebula:a:${definedVersion}'
-                compile 'test.nebula:b:${definedVersion}'
-                compile 'test.nebula:c:${definedVersion}'
-            }
-            """.stripIndent()
-
-        rulesJsonFile << alignTestNebulaRule()
-
-        when:
-        def result = runTasks(*tasks(coreAlignment))
-
-        then:
-        dependencyInsightContains(result.output, "test.nebula:a", ABResultingVersion)
-        dependencyInsightContains(result.output, "test.nebula:b", ABResultingVersion)
-        dependencyInsightContains(result.output, "test.nebula:c", CResultingVersion)
-
-        if (coreAlignment) {
-            // align on the non-failed version
-            assert result.output.contains("belongs to platform aligned-platform:rules-0-for-test.nebula-or-test.nebula.ext:$CResultingVersion")
-        }
-
-        where:
-        definedVersionType | definedVersion | coreAlignment | ABResultingVersion | CResultingVersion
-        "static version"   | "1.4.0"        | false         | "FAILED"           | "1.4.0"
-        "static version"   | "1.4.0"        | true          | "FAILED"           | "1.4.0"
-    }
-
-    @Unroll
-    def 'missing cases: dynamically defined dependency: alignment styles cannot resolve missing dependency when lower versions are missing | core alignment: #coreAlignment'() {
-        given:
-        def mavenrepo = createDepenendenciesForMissingVersionsTests()
-        new File("${projectDir}/gradle.properties").delete()
-
-        buildFile << """
-            repositories {
-                ${mavenrepo.mavenRepositoryBlock}
-            }
-            dependencies {
-                compile 'test.nebula:a:${definedVersion}'
-                compile 'test.nebula:b:${definedVersion}'
-                compile 'test.nebula:c:${definedVersion}'
-            }
-            """.stripIndent()
-
-        rulesJsonFile << alignTestNebulaRule()
-
-        when:
-        def result = runTasks(*tasks(coreAlignment))
-
-        then:
-        dependencyInsightContains(result.output, "test.nebula:a", ABResultingVersion)
-        dependencyInsightContains(result.output, "test.nebula:b", ABResultingVersion)
-        dependencyInsightContains(result.output, "test.nebula:c", CResultingVersion)
-
-        if (coreAlignment) {
-            // align on the non-failed version
-            assert result.output.contains("belongs to platform aligned-platform:rules-0-for-test.nebula-or-test.nebula.ext:$ABResultingVersion")
-        }
-
-        where:
-        definedVersionType | definedVersion | coreAlignment | ABResultingVersion | CResultingVersion
-        "range"            | "0.+"          | false         | "0.5.0"            | "FAILED"
-        "range"            | "0.+"          | true          | "0.5.0"            | "FAILED"
-    }
-
-    @Unroll
-    def 'missing cases: dynamically defined dependency as latest.release: alignment styles attempt to align to highest possible version | core alignment: #coreAlignment'() {
-        given:
-        def mavenrepo = createDepenendenciesForMissingVersionsTests()
-        new File("${projectDir}/gradle.properties").delete()
-
-        buildFile << """
-            repositories {
-                ${mavenrepo.mavenRepositoryBlock}
-            }
-            dependencies {
-                compile 'test.nebula:a:${definedVersion}'
-                compile 'test.nebula:b:${definedVersion}'
-                compile 'test.nebula:c:${definedVersion}'
-            }
-            """.stripIndent()
-
-        rulesJsonFile << alignTestNebulaRule()
-
-        when:
-        def result = runTasks(*tasks(coreAlignment))
-
-        then:
-        dependencyInsightContains(result.output, "test.nebula:a", ABResultingVersion)
-        dependencyInsightContains(result.output, "test.nebula:b", ABResultingVersion)
-        dependencyInsightContains(result.output, "test.nebula:c", CResultingVersion)
-
-        if (coreAlignment) {
-            assert result.output.contains("belongs to platform aligned-platform:rules-0-for-test.nebula-or-test.nebula.ext:$CResultingVersion")
-        }
-
-        where:
-        definedVersionType | definedVersion   | coreAlignment | ABResultingVersion | CResultingVersion
-        "range"            | "latest.release" | false         | "1.1.0"            | "1.4.0"
-        "range"            | "latest.release" | true          | "1.1.0"            | "1.4.0"
-    }
-
-    @Unroll
-    def 'missing cases: dynamically defined dependency as major.+: alignment styles attempt to align to highest possible version | core alignment: #coreAlignment'() {
-        given:
-        def mavenrepo = createDepenendenciesForMissingVersionsTests()
-        new File("${projectDir}/gradle.properties").delete()
-
-        buildFile << """
-            repositories {
-                ${mavenrepo.mavenRepositoryBlock}
-            }
-            dependencies {
-                compile 'test.nebula:a:${definedVersion}'
-                compile 'test.nebula:b:${definedVersion}'
-                compile 'test.nebula:c:${definedVersion}'
-            }
-            """.stripIndent()
-
-        rulesJsonFile << alignTestNebulaRule()
-
-        when:
-        def result = runTasks(*tasks(coreAlignment))
-
-        then:
-        dependencyInsightContains(result.output, "test.nebula:a", ABResultingVersion)
-        dependencyInsightContains(result.output, "test.nebula:b", ABResultingVersion)
-        dependencyInsightContains(result.output, "test.nebula:c", CResultingVersion)
-
-        if (coreAlignment) {
-            assert result.output.contains("belongs to platform aligned-platform:rules-0-for-test.nebula-or-test.nebula.ext:$CResultingVersion")
-        }
-
-        where:
-        definedVersionType | definedVersion | coreAlignment | ABResultingVersion | CResultingVersion
-        "range"            | "1.+"          | false         | "1.1.0"            | "1.4.0"
-        "range"            | "1.+"          | true          | "1.1.0"            | "1.4.0"
-    }
-
-    @Unroll
-    def 'missing cases: dynamically defined dependency as major.minor.+: alignment styles attempt to align to highest possible version | core alignment: #coreAlignment'() {
-        given:
-        def mavenrepo = createDepenendenciesForMissingVersionsTests()
-        new File("${projectDir}/gradle.properties").delete()
-
-        buildFile << """
-            repositories {
-                ${mavenrepo.mavenRepositoryBlock}
-            }
-            dependencies {
-                compile 'test.nebula:a:${definedVersion}'
-                compile 'test.nebula:b:${definedVersion}'
-                compile 'test.nebula:c:${definedVersion}'
-            }
-            """.stripIndent()
-
-        rulesJsonFile << alignTestNebulaRule()
-
-        when:
-        def result = runTasks(*tasks(coreAlignment))
-
-        then:
-        dependencyInsightContains(result.output, "test.nebula:a", ABResultingVersion)
-        dependencyInsightContains(result.output, "test.nebula:b", ABResultingVersion)
-        dependencyInsightContains(result.output, "test.nebula:c", CResultingVersion)
-
-        if (coreAlignment) {
-            assert result.output.contains("belongs to platform aligned-platform:rules-0-for-test.nebula-or-test.nebula.ext:$ABResultingVersion")
-        }
-
-        where:
-        definedVersionType | definedVersion | coreAlignment | ABResultingVersion | CResultingVersion
-        "narrow range"     | "1.0.+"        | false         | "1.0.3"            | "1.0.2"
-        "narrow range"     | "1.0.+"        | true          | "1.0.3"            | "1.0.2"
-    }
-
-    @Unroll
     def 'core alignment uses versions observed during resolution'() {
         // test case from https://github.com/nebula-plugins/gradle-nebula-integration/issues/52
         // higher version transitive aligning parent dependency
@@ -624,65 +259,6 @@ class AlignRulesBasicWithCoreSpec extends IntegrationTestKitSpec {
         ]
     }
 
-    private static String alignSpringRule() {
-        """
-        {
-            "align": [
-                {
-                    "group": "org\\\\.springframework",
-                    "includes": ["spring-(tx|aop|instrument|context-support|beans|jms|test|core|oxm|web|context|expression|aspects|websocket|framework-bom|webmvc|webmvc-portlet|jdbc|orm|instrument-tomcat|messaging)"],
-                    "excludes": [],
-                    "match": "[2-9]\\\\.[0-9]+\\\\.[0-9]+.RELEASE",
-                    "reason": "Align Spring",
-                    "author": "User <user@example.com>",
-                    "date": "2016-05-16"
-                }
-            ]
-        }
-        """.stripIndent()
-    }
-
-    private void springBootBasedBuildFileWith(String extSpringBootVersion, String extSpringVersion, boolean usesForce, String requestedVersion, String additionalPlugin) {
-        buildFile << """
-        buildscript {
-            dependencies {
-                classpath("org.springframework.boot:spring-boot-gradle-plugin:$extSpringBootVersion")
-                classpath "io.spring.gradle:dependency-management-plugin:1.0.7.RELEASE"
-            }
-            repositories {
-                maven {
-                    url "https://plugins.gradle.org/m2/"
-                }
-            }
-        }
-        plugins {
-            id 'java'
-            id 'nebula.resolution-rules'
-        }
-        apply plugin: 'org.springframework.boot'$additionalPlugin
-        repositories {
-            mavenCentral()
-        }
-        dependencies {
-            resolutionRules files('$rulesJsonFile')
-        }
-        ext {
-            springVersion = "$extSpringVersion"
-            springBootVersion = "$extSpringBootVersion"
-        }
-        """.stripIndent()
-
-        if (usesForce) {
-            buildFile << """
-                configurations.all {
-                    resolutionStrategy {
-                        force "org.springframework:spring-core$requestedVersion"
-                    }
-                }
-                """.stripIndent()
-        }
-    }
-
     private static void dependencyInsightContains(String resultOutput, String groupAndName, String resultingVersion) {
         def content = "$groupAndName:.*$resultingVersion\n"
         assert resultOutput.findAll(content).size() >= 1
@@ -703,37 +279,6 @@ class AlignRulesBasicWithCoreSpec extends IntegrationTestKitSpec {
                 ]
             }
         '''.stripIndent()
-    }
-
-    private GradleDependencyGenerator createDepenendenciesForMissingVersionsTests() {
-        def graph = new DependencyGraphBuilder()
-                .addModule('test.nebula:a:0.5.0')
-
-                .addModule('test.nebula:a:1.0.0')
-                .addModule('test.nebula:a:1.0.1')
-                .addModule('test.nebula:a:1.0.2')
-                .addModule('test.nebula:a:1.0.3')
-                .addModule('test.nebula:a:1.1.0')
-
-                .addModule('test.nebula:b:0.5.0')
-
-                .addModule('test.nebula:b:1.0.0')
-                .addModule('test.nebula:b:1.0.1')
-                .addModule('test.nebula:b:1.0.2')
-                .addModule('test.nebula:b:1.0.3')
-                .addModule('test.nebula:b:1.1.0')
-
-                .addModule('test.nebula:c:1.0.0')
-                .addModule('test.nebula:c:1.0.1')
-                .addModule('test.nebula:c:1.0.2')
-                .addModule('test.nebula:c:1.1.0')
-
-                .addModule('test.nebula:c:1.4.0')
-
-                .build()
-        def mavenrepo = new GradleDependencyGenerator(graph, "${projectDir}/testrepogen")
-        mavenrepo.generateTestMavenRepo()
-        return mavenrepo
     }
 
     private GradleDependencyGenerator createDependenciesForExampleAppDependencies() {
